requirements.txt
streamlit
chromadb
google-genai
python-dotenv
sentence-transformers
pysqlite3-binary